perf: use rsync to copy the experiment dir to nodes and undo splitting for nodes into ipv4 and ipv4 address nodes
Undoing the previous change, where we split the nodes into IPv4 and IPv6 addresses, because macOS's rsync wasn't able to sync files to nodes with IPv6 addresses. To work around that issue we create a temporary SSH configuration file with the host and hostname directives.
Essentially, we replace

```sh
$ rsync .. -e 'ssh -o ProxyCommand=".."' <src> host-ipv6:<dst>
```

with

```sh
$ cat <tmp-ssh-config>
Host pegasusvm
  HostName <ipv6-address>
$ rsync .. -e 'ssh -F <tmp-ssh-config> o ProxyCommand=".."' <src> pegasusvm:<dst>
```